Fix scoping of assoc_func in OCSort.cpp and make it useful
Fix scoping for assoc_func, and replace usages of giou_batch with assoc_func to make it actually useful for changing how the code associates tracks and detections. A more in-depth change to the associate function in Association.hpp might be also needed, but would require refactoring the function's arguments and I didn't want to make a breaking change.