Fix performance bug in mutilthread.cpp
The function to update the OCSort tracker was being called once for every detection for every frame. This causes significant slowdowns as well as dropping tracks and hurting Kalman Filter performance. I changed it so that the OCSort tracker update is only called once per frame on the entire vector of detections.